+++ backend/src/main/java/com/javahelp/backend/data/DynamoDBSurveyResponseStore.java
@@ -9,23 +9,16 @@ import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
 import com.amazonaws.services.dynamodbv2.model.ScanRequest;
 import com.amazonaws.services.dynamodbv2.model.ScanResult;
 import com.amazonaws.services.dynamodbv2.model.UpdateItemRequest;
-import com.javahelp.backend.search.constraint.Constraint;
-import com.javahelp.backend.search.constraint.IConstraint;
+import com.javahelp.backend.data.search.constraint.IConstraint;
 import com.javahelp.model.survey.Survey;
 import com.javahelp.model.survey.SurveyQuestion;
 import com.javahelp.model.survey.SurveyQuestionResponse;
 import com.javahelp.model.survey.SurveyResponse;
-import com.javahelp.model.user.User;
-import com.javahelp.model.user.UserPassword;
-
-import org.w3c.dom.Attr;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
