+++ backend/src/main/java/com/javahelp/backend/data/ISurveyResponseStore.java
@@ -1,7 +1,7 @@
 package com.javahelp.backend.data;
 
 import com.amazonaws.regions.Regions;
-import com.javahelp.backend.search.constraint.IConstraint;
+import com.javahelp.backend.data.search.constraint.IConstraint;
 import com.javahelp.model.survey.SurveyResponse;
 import com.javahelp.model.user.User;
