+++ backend/src/main/java/com/javahelp/backend/data/search/constraint/Constraint.java
@@ -1,19 +1,12 @@
-package com.javahelp.backend.search.constraint;
+package com.javahelp.backend.data.search.constraint;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
-public class MultiOptionConstraint implements IConstraint {
+public class Constraint implements IConstraint {
     private Set<String> constraints = new HashSet<>();
 
-    public MultiOptionConstraint() {}
-
-    public void setMultiConstraint(List<String> constraintList) {
-        for (String constraint : constraintList) {
-            setConstraint(constraint);
-        }
-    }
+    public Constraint() {}
 
     @Override
     public void setConstraint(String constraint) {
@@ -23,7 +16,7 @@ public class MultiOptionConstraint implements IConstraint {
     @Override
     public Set<String> getConstraints(){
         return this.constraints;
-    }
+    };
 
     @Override
     public int size() { return this.constraints.size(); }
